Enable the drm_hwcomposer and minigbm gralloc modules.
These modules are enabled in addition to the vsoc implementations for
now, so we can run with the old and new models for graphics and switch
between them using device properties. This change does nothing except
turn on the modules; they are not used in any way.

Bug: 77276633
Change-Id: I88bb24d3cab1b4b3cd14b98ced0e84d61e538def
Merged-In: I88bb24d3cab1b4b3cd14b98ced0e84d61e538def