Remove demo HVAC property in favor of using the emulated VHAL.
The use of this feature is redundant with the emulated VHAL. It is also HVAC specific which does not align well with the future design of accessing car properties. When testing/mocking for a specific property is needed, it is preferable that this is injected at a lower level.

Bug: 110105682

Test: build and deploy
Change-Id: Ie2fc212984e66eadd9d72a193493e4e0d7e7d972
(cherry picked from commit ebc892b37955a11574824cc8221abfc724f1833e)