Make it possible to disable SYSTEM_ROOT_IMAGE feature
In newer platform versions, the system-as-root image feature is
deprecated, eliminating the need to pass the Android fstab through
device-tree. As cuttlefish only uses the device-tree for the Android
fstab, we can stop using it altogether.

For maximum backwards compatibility, this change does not actually
remove the build/install of the initrd-root.dtb file, but the file is
not required any more. An fstab.vsoc file is installed in the root of
the initramfs which describes all of the mount points.

This change can be enabled with TARGET_BUILD_SYSTEM_ROOT_IMAGE=false
on the build command line.

This change also makes it possible to boot newer platform versions on
hypervisors that do not support device-tree.

Bug: 120163151
Test: m -j128 TARGET_BUILD_SYSTEM_ROOT_IMAGE=false dist && launch_cvd
Change-Id: I507b7baa7b65908dd0e877ccc086c61b4026028a