Remove first_stage_mount flag from metadata
Until the DAP conversion is merged, we're still passing a dtb to the
vm_manager, so we have to split the dtb and fstab to avoid double
mounts. We can't put metadata in the dts/dtb fstab either, because it
might not be formatted during first_stage_init, and nothing is there to
format it.

For DAP builds, this is not a problem, because although the metadata
mount might fail, we install the same fstab to ramdisk and vendor, and
fsmgr will handle this for us.

Bug: 128222953
Change-Id: Ib275611fb9b48de038803a80c2e4ab1081e9f19b