Remove obsolete gpu host libraries
libvirglrenderer is built as part of the crosvm build process, so we
don't want to package it here.

We've decided against landing SwiftShader on the host side for the time
being, so there's really no need to build it and put it into the host
package.

Test: Build and boot
Change-Id: I85b60d0c1a47ae8b433c135c653598e669749c8d
(cherry picked from commit 525e288fda7fcc7ab6288e0ed1f2265d50323ef4)