Support composite disks.
This adds an extra vendor_init property "ro.boot.fstab_name" which
controls whether the composite or multi-device fstab is used.

With the extra file_contexts and the new variable it's possible to boot
the same partition contents either separated out to different drives or
composited to a single drive.

Bug: 133432409
Test: launch_cvd -vm_manager=crosvm
Test: launch_cvd -vm_manager=crosvm --composite_disk=composite.img
Change-Id: I2503ec27b739fe26e3eb1c4ae777df8ef5ee573c