Convert pasan build to aosp_cf schema
This config is supposed to be the same as phone, just with a different
name, as a workaround for an android-build issue. It was mistakenly not
converted. Convert it now.

Bug: 112341596
Change-Id: I7631da6f23e41637938761660c60413c714ef1cd
Signed-off-by: Alistair Delva <adelva@google.com>
Merged-In: I7631da6f23e41637938761660c60413c714ef1cd
(cherry picked from commit aaabb2c606661c22485a50632c7c45c49ffab4c8)