Add support for DFU over TFTP to ARM host image
This adds support for firmware upgrades over TFTP.
The device will broadcast BOOTP messages on the
network to find the TFTP server.  If one is running,
it attempts to download the manifest in the expected
format.  The manifest has a SHA field that describes
the contents of the update.  If this is different
than the current version, the device will run through
the rest of the manifest and request those image files
from the TFTP server.  Gzipped files are supported.

BUG: 142283345
Test: local build and run on rock pi 4b
Signed-off-by: Tristan Muntsinger <muntsinger@google.com>
Change-Id: I3e9756a03ad30aff33df6f6329e0e41a893cfa4d
(cherry picked from commit bcdc07b903a1803c9443ead275772cf4521dedc8)