Run qemu without vsoc
This enables the crosvm options on qemu and disables references to the
vsoc driver. The vsoc driver was specially written to share memory with
the guest and the host through ivsh, but it is not included in the 4.19
kernel due to difficulty including it as a module in the gki setup.

Test: launch_cvd -vm_manager=qemu_cli, check adb and vnc
Bug: 143713267
Change-Id: I747f7f4c2f21c2e737899a07f1d8e8439da3ea6e