Extract the target zip's system image for old devices
As pie-gsi builds are unaware of the super image, this will extract the
system image out of the target files zip to maintain the drop-in
functionality of older branches.

pie-gsi appears to conflict with later devices about setting some
variables on the tombstone monitor, that will have to be solved
separately.

Test: Mix aosp-pie-gsi with aosp-android10-gsi, aosp-master
Bug: 145829296
Change-Id: I34d7746934dfdc7cdb5dd26cabde51e88f461f60