Make subprocesses die when their parents do
This could also be a configurable option in cvd::Subprocess, but it
seems to work when always enabled.

The PR_SET_PDEATHSIG setting is unset on the forked child, which happens
to do the right thing already with run_cvd daemonizing. Because that does
not use cvd::Subprocess, the setting doesn't apply to it (though it is
given to run_cvd originally by launch_cvd).

Bug: 145956571
Test: ctrl-c on launch_cvd
Change-Id: I387db1acabe90e1f41abb159b1c07e891ea46a87