Use big enough types to accomodate port numbers
GetVsockPort() returns an unsigned int, so that's the type that should
be used everywhere. The one exception is when obtaining it from a
property in the guest, where unsigned int is not available, in which
case int64 is used instead.

Bug: 146582083
Test: launch_cvd and get lucky with a big port
Change-Id: I4738a7a07dd41d0d72d9b4d1628ac8fe9ef328b9
(cherry picked from commit 36845e3f05a0d6e744a64783ad6344e6a5203301)