Reduce TCP socket load for single send
The size of the payload for one socket send is unbounded. This leads
to the error "Resource is unavailable" when the socket cannot handle
the load, which is caught as FATAL and thereby crashes the host side.

We break the payload up into smaller parts.

Bug: 152441545
Test: navigate around to induce low-compression rate frames
      (see bug for description) and not crash
Change-Id: Ic1147d227b7a5cf91c1bfb42bd7fd1eec59447d1
(cherry picked from commit aef0a2dba653f4fc75b919ad71d53b6121cd4620)