Move reference-ril to cuttlefish
To verify the modem simulator, reference-ril needs to be modified a lot,
so move it here. And based on this commit, make some modifications to verify
modem simulator and pass vts, cts.
It's easier to see what has been modified based on this commit.

Note: it is the same as go/oag/1294202/
but with different change id to break cl dependency
so to merge one by one.

Bug: 152799213
Change-Id: I3ba2cbec9046ca175fdebc50e9e0dcd00edc3122
(cherry picked from commit 61a8030cbdf4fd0e2872dc538a2af0e9ed8a8b0d)