Re-order bootloader partitions in the composite disk
Fix up the a/b slotting for boot and move misc to the top of the
partition table. Move all of the partitions needed by the bootloader
to the top of the disk so they are less likely to be reordered.

This also fixes a pie-gsi specific problem where the a/b block devices
were not properly labelled by sepolicy.

Bug: 155019925
Change-Id: Ic77fd0cdf8d28ea533610773428a4dbbcd368095