Enable chained vbmeta on cuttlefish
Remove a mainline override that disabled AVB. This change also enables
chained vbmeta so that system/system_ext/vbmeta_system can be replaced
independently of the rest of the system, which will be useful for future
build mixing.

This change works around a "feature" of libavb which always reads the
maximum vbmeta payload size of 64KB, even though the build system aligns
the vbmeta/vbmeta_system partitions to 4KB, and on cuttlefish they are
8K/4K respectively.

Finally, enable --flags 2 for now so that this new functionality is
disabled at runtime until we can provide a valid value for
'androidboot.vbmeta.digest'.

Bug: 155019925
Change-Id: Ie0fd97fd8907e5cc43051c2385be18242e6d3714
(cherry picked from commit 56959d806ddf5542f548d87866523d4bea5a2944)