[WebRTC] Add the adb endpoints to the client library
It allows to exchange data directly between the client (browser) and
the adb daemon running in the guest. It's only used to show logcat
ouput for now as an example.

The current implementation temporarily uses the signaling server to
forward the data, but that will migrate to using a direct data channel
between client and device when possible.

Bug: 155122564
Test: locally
Change-Id: I5276ed070c4baa3be42a54096911032a4967021a