Detect misconfigured tap/bridge interfaces
If the mobile or wireless bridges or tap devices are incorrectly
configured, without a broadcast address, the cuttlefish RIL setup code
will incorrectly derive a bad broadcast address and pass it to the
guest. This sets up a bad routing table which causes corruption of the
MAC address for packets routed out, which makes connectivity impossible.

Detect this strange condition and print out an error, instead of
allowing the device to continue to be set up with the bad address.

Bug: 156103265
Change-Id: I24a70c45c4468e7db7fd128de00acd4b7b7ae407
(cherry picked from commit ff364b7576cec562af28fb6b49777bda1ad6f9e7)