Don't destruct Packetizer when RTPSender is gone
Cherry-picked from qt-stratus-dev

Currently, references to the packetizer are only held by
RTPSender. When all RTPSenders that are connected to a packetizer are
destructed, the packetizer will also be destructed. When new connection
comes in, a new packetizer is created, and the new packetizer calls
FrameBufferSource.start() which resets the VPXEncoder in the
FrameBufferSource.

This leads to two problems:
1. The mNumConsumers count in FrameBufferSource is inaccurate as the
packetizer is destructed before it gets a chance to notify the
FrameBufferSource.
2. The storedFrame of VPXEncoder gets reset. As a result, a new
connection has to wait for a new frame from wayland.

In the current model, there is a 1:1 relation between FrameBufferSource
and Packetizer. Hence, packetizer should not be destroyed when there is
no client.

Test: manual
Bug: 150979403
Change-Id: I12fb019d336acc22d9e891abe4e156749cab539e