Remove touchscreen/faketouch from tv
Enabling touchscreen is wrong for TV because is will cause the wrong
version of gboard to be pushed by play services to the device.

This change may break VNC or WebRTC interactions with the TV build, so
use a physical controller or the ATV remote app to work around it.

Bug: 158883251
Change-Id: Ibb85587c1b04a474945c268a38b4fc873ffbb015
(cherry picked from commit 77bec7c7093443c294ed9ed203d8545884bcd207)