Fix QEMU 5.0 aarch64 boot
An old romfile bug has been fixed, and now the rom must be specified
with -optionrom, so our obsolete workaround is a syntax error.

The gic-version flag is also incorrect, it uses a hyphen instead of
an underscore, and the underscore is a syntax error.

Bug: 161538883
Change-Id: I811647b4911eea10c93585ac58e43247955487db
(cherry picked from commit cb021af0bd9d090cea5bc8cc2c2ec13f14741edf)