Fix cuttlefish_libicuuc defaults
For host components, libicuuc must be linked explicitly, however for
target it will be inherited from shared_libs. The androidicu library
isn't required on the host, and is wrong for target components as it
is not available to 'vendor', so remove it.

Bug: 167675429
Change-Id: I76536b01c77bad74ac9604cec2f504fa797037ff
(cherry picked from commit 8cd279fa9a040d99b690e14e44021ef7c8c131f8)