Remove most of cuttlefish_time
This is only used by the hwcomposer for stats generation, and only the
interface provided by the C++ header is used, not the library. Drop
the library for now but leave the header, and we can replace with
std::chrono in the future.

Bug: 167675429
Bug: 168244215
Change-Id: I19633640f2fc9d2fce30e8f14eeca105a4ea9d77
(cherry picked from commit ab1726bfacaf1144897980724bc2f48c3c3ac05e)