Rename cuttlefish properties to vendor_
Keep VtsTrebleSysProp happy.

Remove the "cf_" or "cuttlefish_" prefix if it is not present in the
property name; "vendor_" is sufficient. Renamed some of the properties
for better consistency (base on the property name itself, unless it is
a property group, then give it a descriptive name).

Bug: 157958356
Bug: 173683489
Change-Id: Icd4942037d94f8bf17d0712f3c67623625c87576
(cherry picked from commit 207b258d73c436428b10454279600644419cd516)