Fix -console=true with QEMU
The serial console code had bitrotted and it was still using a socket
instead of a pty, which was very confusing. Update it to use a pipe
and rely on the newly refactored ConsoleForwarder. QEMU does support
'pty' natively, but it does not allow the pty path to be specified
or retrieved, so it's not very usable.

Change-Id: I92dbbe9b1ac77d38a42d7640d608094e4dfee959
(cherry picked from commit 44f9698da627e38830d21f7a412427e82ab17888)