Cache first frames before VNC client connection
A VNC session sometimes starts with black screen. Cuttlefish waits for
a vnc client connection first before requesting a frame from the guest.
The guest might respond slowly when it thinks it's already in the
steady state. This makes cuttlefish VNC server wait for a while to
have the first frame. This fix fetches Android guest frames early on
and caches it in the queue consumed by the cuttlefish VNC server.

Bug: b/184740490
Test: launch_cvd --start_vnc_server locally
Change-Id: Ib3c76331f36506c2b8785ebc7ed1ed5e01ba7bfd