Create the peer connection from the client handler
Part of the creation of the connection still happens in the streamer
but it's called from the client handler during handling of the
'request-offer' message. This is needed to be able to add the ice
servers coming in that message to the peer connection.

Bug: 228862302
Test: locally
Change-Id: I936b217820d37c38e659f153b46d779655b9ac97
(cherry picked from commit 87ceb8ecb7cd14f79b50422f0fff4c6108aec757)