Move Gatekeeper, SecureEnv K. Argument Generation
earlier in launch_cvd to assemble_cvd. This change is needed to
finish kernel argument generation and save that text to the
environment partition which is needed for uboot to
startup.

This commit is generated partially from af4b621f3 and c2bddd1.

Bug: 153909141
Bug: 181351700
Test: Treehugger
Change-Id: I71f859ff28da4dbe77a6ecfb30e9fc39f7116242
Merged-In: I71f859ff28da4dbe77a6ecfb30e9fc39f7116242