Add generic blocklist for vkms
When vkms was added to android13-5.15 and later kernels, a blocklist was
added to prevent the module from loading all the time, so a second DRM
display would not be created. However, no platform-side blocklist file
was added, so when using the checked-in kernel prebuilt, the vkms module
was still loaded.

This caused problems with -gpu_mode=drm_virgl because the Mesa driver
would probe /dev/dri/card*, but sepolicy was not written for card1, so
SELinux denials would be seen.

Add a platform blocklist so no kernel with vkms will have that driver
loaded. The user is expected to load this module manually if booting
with virtio_gpu disabled, for testing only.

Change-Id: Idbc839022a1b67d04222425c3b080f3c2a812ef0