Rework wakeup SELinux policy for ACPI and block virtio PCI nodes
Currently genfscon is used to set sysfs_wakeup context for sysfs files
of ACPI and block devices associated with virtio PCI devices. Since
genfscon only supports path prefixes it requires adding entries for
each device.

This means that we need to update the SELinux policy everytime a new
virtio device is added which is easily missed and is hard to maintain.

This patch rework the SELinux policy to use file_contexts which
supports full regular expressions. This simplifies the policy as there
are now only two rules for ACPI devices (one for crosvm and one for
qemu) instead of the few tens before. It also simplifies the process
of adding new virtio devices as now the SELinux policy does not need
to be updated with each new virtio device.

Bug: 307608726
Test: check no avc denial regressions at boot for both qemu (x86, arm64, riscv64) and crosvm (x86)
Change-Id: Iec1492420aae55ed3abfd4315d55f7a70191c135