Remove dirs and files across mounts in containers
When running in a container, the root filesystem uses OverlayFS. Due to
a quirk of the OverlayFS implementation, nftw() will not include certain
files in the walk since their device number differs from the parent
directory. We remove the FTW_MOUNT flag to ensure that all child file and
directories are deleted before deleting the parent.

Bug: 465804122
Test: Run target in a container, run "cvd create" and "cvd rm". Ensure
      no errors mentioning RecursivelyRemoveDirectories are logged.
Change-Id: Iad1cfc4b5feaa0c860e6eb411db857e0f600a10f