run_cvd exits immediately on SIGINT
When running in daemon mode run_cvd doesn't receive the interrupt. It
only exited after the boot completes, it tries to write the exit code to
the foreground launcher and receives a SIGPIPE. By checking the pipe to
the foreground launcher on the boot state machine loop it can notice the
foreground launcher terminating early and terminate itself.

Bug: b/467794263
Test: locally
Flag: EXEMPT cuttlefish host code
Change-Id: I030528b5aeda7028d4b0e7bd9955b754d4a6b085